@@ -69,7 +69,16 @@
     <!-- Content (Item / Catalog) -->
     <router-view />
     <footer>
-      <small class="poweredby text-muted" v-html="poweredByText" />
+      <ul v-if="Array.isArray(footerLinksFromVueX) && footerLinksFromVueX.length > 0" class="footer-links text-muted">
+        <li v-for="link in footerLinksFromVueX" :key="link.url">
+          <a :href="link.url" target="_blank">{{ $te(`footerLinks.${link.label}`) ? $t(`footerLinks.${link.label}`) : link.label }}</a>
+        </li>
+      </ul>
+      <i18n-t tag="small" keypath="poweredBy" class="poweredby text-muted" scope="global">
+        <template #link>
+          <a href="https://github.com/radiantearth/stac-browser" target="_blank">STAC Browser</a> {{ browserVersion }}
+        </template>
+      </i18n-t>
     </footer>
     <b-popover
       v-if="root" id="popover-root" class="popover-large" target="popover-root-btn"
@@ -150,6 +159,7 @@ export default defineComponent({
     ...mapState(['allowSelectCatalog', 'conformsTo', 'data', 'dataLanguage', 'globalError', 'loading', 'stateQueryParameters', 'uiLanguage', 'url']),
     ...mapState({
       catalogImageFromVueX: 'catalogImage',
+      footerLinksFromVueX: 'footerLinks',
       localeFromVueX: 'locale',
       detectLocaleFromBrowserFromVueX: 'detectLocaleFromBrowser',
       supportedLocalesFromVueX: 'supportedLocales',
@@ -165,10 +175,6 @@ export default defineComponent({
       else {
         return "";
       }
-    },
-    poweredByText() {
-      const link = `<a href="https://github.com/radiantearth/stac-browser" target="_blank">STAC Browser</a> ${this.browserVersion}`;
-      return this.$t('poweredBy', { link });
     },
     isSearchPage() {
       return this.$route.name === 'search';
@@ -217,13 +223,13 @@ export default defineComponent({
     type() {
       if (this.data instanceof STAC) {
         if (this.data.isItem()) {
-          return this.$tc('stacItem');
+          return this.$t('stacItem', 1);
         }
         else if (this.data.isCollection()) {
-          return this.$tc(`stacCollection`);
+          return this.$t(`stacCollection`, 1);
         }
         else if (this.data.isCatalog()) {
-          return this.$tc(`stacCatalog`);
+          return this.$t(`stacCatalog`, 1);
         }
         else if (Utils.hasText(this.data.type)) {
           return this.data.type;
