@@ -1,6 +1,6 @@
 <template>
   <section v-if="collection" class="parent-collection card-list mb-4">
-    <h2>{{ $tc('stacCollection') }}</h2>
+    <h2>{{ $t('stacCollection', 1) }}</h2>
     <Catalog :catalog="collection" :showThumbnail="showThumbnail" />
   </section>
 </template>
